Widget up and running. Closed #9
Widget reflects the current status of the light. Tap the widget to
toggle the status. The change is reflected in the image of the widget.
The widget is also scaleable. There is a bit of a hack in how i hold
onto the AppWidgetManager but I couldn't find a way to do it that didn't
involve throwing it around(and the list of widgetIds) to the
asyncAPICall class, which I really didn't want to to do as it adds code
complexity for a single tiny use case.

Also added a GetStatusTask, which simply returns the current light
status from the Spark Core.