Observer pattern for toggle. Worked on spark lib
Sparklib supports login and device finding.
Added observer pattern to the toggle task so that all observers are
updated upon a change in the status of the bulb. This fixes the issue
where toggling status in the app does not update the widget, and
vice-versa. As far as spark lib goes, i’ve added Token and Device
classes, which hold more data than the previous strings.