Closed #25
Displays a toast message on the timeout. Takes 8-10 seconds for the
device to time out. I'll actually try to shorten it to 3-4 if
possible... otherwise we'll have to implement internal timers to kill it
after a few seconds and i really dont wanna do that.
Toast message works in: Widget, Gesture, Application Button.